+++ runtime-light/tl/tl.cmake
@@ -1,1 +1,1 @@
-prepend(RUNTIME_LIGHT_TL_SRC tl/ tl-builtins.cpp tl-functions.cpp tl-types.cpp)
+prepend(RUNTIME_LIGHT_TL_SRC tl/ tl-functions.cpp tl-types.cpp)
